requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
nats-py==2.11.0
pytest==8.3.3
httpx==0.27.2
python-dotenv==1.0.1
asyncpg==0.29.0
